override mid level with refactor
