+++ torch_concepts/__init__.py
@@ -9,8 +9,8 @@ from typing import Any
 
 from .annotations import Annotations, AxisAnnotation
 from .nn.modules.utils import GroupConfig
-from .nn.modules.mid.constructors.concept_graph import ConceptGraph
-from .nn.modules.mid.models.variable import Variable, LatentVariable, InputVariable, ExogenousVariable, ConceptVariable, EndogenousVariable
+from .nn.modules.concept_graph import ConceptGraph
+from .nn.modules.mid.models.variable import Variable, ConceptVariable, EmbeddingVariable
 from .utils import seed_everything
 from . import nn, distributions
 from . import data
@@ -34,11 +34,8 @@ __all__ = [
 
     # Variables
     "Variable",
-    "LatentVariable",
-    "InputVariable",  # Backward compatibility alias
-    "ExogenousVariable",
     "ConceptVariable",
-    "EndogenousVariable",  # Backward compatibility alias
+    "EmbeddingVariable",
 
     "seed_everything",
