+++ torch_concepts/nn/modules/mid/__init__.py
@@ -1,5 +1,9 @@
-"""
-Mid-level API for torch_concepts.
-"""
+"""Mid-level package (PGM-based CBMs) — internal.
 
+The public API is re-exported from :mod:`torch_concepts.nn`; import mid-level
+symbols from there rather than from this package.
+
+.. warning::
+    Experimental APIs subject to change without a deprecation period.
+"""
 __all__: list[str] = []
