--- torch_concepts/nn/modules/mid/base/__init__.py
@@ -1,1 +1,0 @@
-__all__: list[str] = []
