+++ torch_concepts/nn/modules/mid/constructors/graph.py
@@ -3,11 +3,11 @@ from torch.nn import Identity, Module
 
 from .....annotations import Annotations
 from ..models.variable import Variable, LatentVariable, ExogenousVariable, ConceptVariable
-from .concept_graph import ConceptGraph
+from ...concept_graph import ConceptGraph
 from ..models.cpd import ParametricCPD
 from ..models.probabilistic_model import ProbabilisticModel
 from .....distributions import Delta
-from ..base.model import BaseConstructor
+from .model import BaseConstructor
 from ...low.lazy import LazyConstructor
