+++ torch_concepts/nn/modules/mid/constructors/model.py
