+++ torch_concepts/nn/modules/mid/inference/pyro/__init__.py
@@ -1,0 +1,2 @@
+"""Pyro-backed inference engines (internal). Public API: :mod:`torch_concepts.nn`."""
+__all__: list[str] = []
