+++ torch_concepts/nn/modules/mid/inference/torch/__init__.py
@@ -1,0 +1,2 @@
+"""Pure-PyTorch inference engines (internal). Public API: :mod:`torch_concepts.nn`."""
+__all__: list[str] = []
