+++ torch_concepts/nn/modules/mid/models/__init__.py
@@ -1,1 +1,2 @@
+"""Mid-level model components (internal). Public API: :mod:`torch_concepts.nn`."""
 __all__: list[str] = []
