--- torch_concepts/nn/modules/pgm/inference/pyro/__init__.py
@@ -1,11 +1,0 @@
-"""Pyro-backed inference engines."""
-from .base import PyroBaseInference
-from .variational import PyroVariationalInference
-from .utils import dist_to_params, trace_to_params
-
-__all__ = [
-    "PyroBaseInference",
-    "PyroVariationalInference",
-    "dist_to_params",
-    "trace_to_params",
-]
