--- torch_concepts/nn/modules/pgm/inference/torch/__init__.py
@@ -1,14 +1,0 @@
-"""Pure-PyTorch inference engines (no Pyro dependency)."""
-from .base import TorchBaseInference
-from .forward import TorchForwardInference
-from .deterministic import TorchDeterministicInference
-from .ancestral import TorchAncestralInference
-from .rejection import TorchRejectionSampling
-
-__all__ = [
-    "TorchBaseInference",
-    "TorchForwardInference",
-    "TorchDeterministicInference",
-    "TorchAncestralInference",
-    "TorchRejectionSampling",
-]
