--- torch_concepts/nn/modules/pgm/models/__init__.py
@@ -1,23 +1,0 @@
-from .variable import (
-    ConceptVariable,
-    EmbeddingVariable,
-    Variable,
-    Delta,
-    PARAM_DIM,
-)
-from .factor import ParametricFactor
-from .cpd import ParametricCPD
-from .probabilistic_model import ProbabilisticModel
-from .bayesian_network import BayesianNetwork
-
-__all__ = [
-    "Variable",
-    "ConceptVariable",
-    "EmbeddingVariable",
-    "Delta",
-    "PARAM_DIM",
-    "ParametricFactor",
-    "ParametricCPD",
-    "BayesianNetwork",
-    "ProbabilisticModel",
-]
